Allow absolute redirect URIs
Currently, when you configure a `redirect_uri` that is an absolute URI, it has
its path extracted and appended to the request domain to create the actual
`redirect_uri` sent to Github. So, for example, if you have a `redirect_uri` of
`test1.example.com/callback` but your app runs on `test2.example.com`, the
`redirect_uri` sent to Github will be `test2.example.com/callback`. In most
cases, this is the intended behavior, since Github doesn't allow redirect URIs
that aren't extensions of the callback URI.

This patch allows `redirect_uri`s that are specified as absolute URIs to
retain their domains, which allows, for example, OAuth schemes where a single
app is the "OAuth broker" for several other apps. See
http://stackoverflow.com/a/13841857 for an example. Using this patch, I was able
to set up a single OAuth broker for many apps as described in the example.