Improve JSON serialization compatibility of User
Previously, the MembershipCache was a subclass of Hash. When serializing
to and deserializing from JSON, the MembershipCache would become a Hash
and blow up. This is a problem in warden-github-rails where I switched
to a custom warden (de)serialization in order to properly support JSON
serialization and also to support mocked users in test environments
between requests[1].

[1]: https://github.com/fphilipe/warden-github-rails/commit/db82470e6fa66b0b546cfef501abcd04f91529ca