Persist membership cache between requests
The membership cache was being serialized into the user object. It was
assumed that the user object was serialized at the end of each request.
It turns out that this is not the case. Instead, the user object only
gets serialized once, namely after setting the user. This means that the
membership cache wasn't working at all between requests.

This commit fixes it by not including the memberships in the serialized
user as there's no point in doing so. Instead, after the user object is
set up, the memberships hash get retrieved from the user session (warden
provides a session object for each scope; if there is none, a new hash
is stored in the session) and assigned to the user object. The hash is
then used as before and we get the session serialization from rack for
free.

Fixes #52.